For Game.c, I noticed that UNI_B was being called UNI_A at initialise. While inconsequential, it'd be best to correct that.
For mechanicalTurk.c, I basically copied a lot of our Game.c code to that. It reads all the game data, so we can basically access it with d instead of g. This will let us easily think about writing an AI, and minimises the amount of "interface crossing".

We will need to write a function that converts a co-ord (or two for edges) into a path. The path needs to figure out where to go to reach the destination, and translate that into letters.